+++ server/package.json
@@ -18,7 +18,6 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
-    "passport": "^0.5.3",
     "passport-google-oauth20": "^2.0.0",
     "server": "file:",
     "sqlite": "^5.1.1",
