+++ server/src/initDB.ts
@@ -31,11 +31,9 @@ const initDB = async () => {
                 email TEXT NOT NULL,
                 calendar_day TEXT NOT NULL,
                 time_start TEXT,
-                time_end TEXT,
                 time_slept TEXT,
                 checklist TEXT,
-                desired_bedtime TEXT,
-                desired_reminder_time TEXT
+                bedtime TEXT,
             );
         `);
         
@@ -54,7 +52,7 @@ const initDB = async () => {
 
         // Insert test data into `calendar_entries` table
         await db.run(`
-            INSERT OR IGNORE INTO calendar_entries (id, email, calendar_day, time_start, time_end, time_slept, checklist, desired_bedtime, desired_reminder_time)
+            INSERT OR IGNORE INTO calendar_entries (id, email, calendar_day, time_start, time_slept, checklist, bedtime)
             VALUES 
                 ('1', 'testuser@example.com', '2024-11-20', '22:00', '06:00', '8', '{"task1": "done", "task2": "pending"}', '22:00', '21:30'),
                 ('2', 'exampleuser@example.com', '2024-11-21', '23:00', '07:00', '8', '{"task1": "completed", "task2": "pending"}', '23:00', '22:30');
